Statement-2-Lucknow-Foodie/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
chromadb
google-generativeai
requests
pydantic
python-multipart
